backend/requirements.txt
# RigidLabeler Backend Dependencies

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation
pydantic>=2.0.0

# Configuration
pyyaml>=6.0

# Numerical computation
numpy>=1.24.0

# Image processing (optional, for warp preview)
opencv-python>=4.8.0
Pillow>=10.0.0

# Development dependencies
# pytest>=7.4.0
# httpx>=0.25.0  # For testing FastAPI
